Entrega final del proyecto

+++ controllers/usersController.js
@@ -4,6 +4,7 @@ const bcrypt = require('bcryptjs');
 const fs = require('fs');
 const db = require('../database/models');
 const Op = db.Sequelize.Op;
+let provinces = ['Buenos Aires', 'Catamarca', 'Chaco', 'Chubut', 'Corrientes', 'Entre Rios', ' Formosa', 'Jujuy', 'La Pampa', 'La Rioja', 'Mendoza', 'Misiones', 'Neuquen', 'Rio Negro', 'Salta', 'San Juan', 'San Luis', 'Santa Cruz', 'Santa Fe', 'Santiago del Estero', 'Tierra del Fuego', 'Tucuman']
 
 const usersController = {
     login: (req, res) => {
@@ -13,10 +14,15 @@ const usersController = {
     loginProcess: async (req, res) => {
         // volcamos valores del form en variables
         const { username, password, recordarUsuario } = req.body;
+        let userToLogin;
+        try {
+            userToLogin = await db.User.findOne({
+                where: { email: username },
+            });
+        } catch (error) {
+            throw new Error('Hubo un error al conectarse a la db');
+        }
 
-        const userToLogin = await db.User.findOne({
-            where: { email: username },
-        });
         // return res.send(userToLogin);
         if (userToLogin) {
             const pwdMatch = bcrypt.compareSync(password, userToLogin.password);
@@ -43,6 +49,15 @@ const usersController = {
                     },
                 });
             }
+        } else if (!userToLogin && username.trim() === '') {
+            return res.render('users/login', {
+                title: 'Login',
+                errors: {
+                    email: {
+                        msg: 'Ingresa tu nombre de usuario',
+                    },
+                },
+            });
         } else {
             return res.render('users/login', {
                 title: 'Login',
@@ -56,47 +71,69 @@ const usersController = {
     },
 
     register: (req, res) => {
-        return res.render('users/register', { title: 'Register' });
+        return res.render('users/register', { title: 'Register', provinces});
     },
 
     processRegister: async (req, res) => {
+        
         // renderizado de errores
+        console.log(req.body);
         const results = validationResult(req);
         if (results.errors.length > 0) {
+            if (req.file?.filename) {
+                fs.unlinkSync(`./public/images/users/${req.file.filename}`);
+            }
             return res.render('users/register', {
                 title: 'Register',
+                provinces,
                 errors: results.mapped(),
                 oldData: req.body,
             });
         }
         // validacion de que el email utilizado para el registro no esta en uso
-        const userInDb = await db.User.findAll({
-            where: { email: req.body.email },
-        });
-        // return res.send(userInDb);
-        if (userInDb.length > 0) {
-            return res.render('users/register', {
-                title: 'Register',
-                errors: { email: { msg: 'Este email ya esta registrado' } },
-                oldData: req.body,
+        try {
+            const userInDb = await db.User.findAll({
+                where: { email: req.body.email },
             });
+            if (userInDb.length > 0) {
+                return res.render('users/register', {
+                    title: 'Register',
+                    errors: { email: { msg: 'Este email ya esta registrado' } },
+                    oldData: req.body,
+                });
+            }
+
+            // creacion de usuario
+            const userToCreate = {
+                ...req.body,
+                password: bcrypt.hashSync(req.body.password, 10),
+                avatar: req.file?.filename ? req.file.filename : 'default.jpg',
+            };
+            // agregar user al archivo users.jsons
+            db.User.create(userToCreate)
+                // Setamos el campo 'detail' para luego pasarlo via API
+                .then((user) => {
+                    db.User.update(
+                        {
+                            detail: `/api/users/${user.id}`,
+                        },
+                        {
+                            where: { id: user.id },
+                        }
+                    );
+                    return res.redirect('/users/login');
+                })
+                .catch((err) => {
+                    throw new Error(
+                        'occurio un error durante la creacion del usuario, intente mas tarde'
+                    );
+                });
+        } catch (error) {
+            throw new Error('Ocurrio un error al validar email');
         }
-
-        // creacion de usuario
-        const userToCreate = {
-            ...req.body,
-            password: bcrypt.hashSync(req.body.password, 10),
-            avatar: req.file.filename,
-        };
-        // agregar user al archivo users.jsons
-        db.User.create(userToCreate);
-
-        // return res.send('se guardo el usuario');
-        return res.redirect('/users/login');
     },
 
     userProfile: (req, res) => {
-        // return res.send(req.session.userLogged);
         return res.render('users/userProfile', {
             title: 'Profile',
             user: req.session.userLogged,
@@ -129,30 +166,42 @@ const usersController = {
         } else {
             // Almacenamos los campos editados
             // Si el usuario cambia la imagen, borramos la anterior
-            if (req.file) {
-                fs.unlinkSync('./public/images/users/' + userLogged.avatar);
-            }
+            // if (req.file) {
+            //     fs.unlinkSync('./public/images/users/' + userLogged.avatar);
+            // }
             // Actualizamos la base de datos
-            await db.User.update(
-                {
-                    first_name: userEdits.first_name,
-                    last_name: userEdits.last_name,
-                    email: userEdits.email,
-                    avatar: req.file ? req.file.filename : userLogged.avatar,
-                },
-                {
-                    where: { id: req.params.id },
-                }
-            );
+            try {
+                await db.User.update(
+                    {
+                        first_name: userEdits.first_name,
+                        last_name: userEdits.last_name,
+                        email: userEdits.email,
+                        avatar: req.file
+                            ? req.file.filename
+                            : userLogged.avatar,
+                    },
+                    {
+                        where: { id: req.params.id },
+                    }
+                );
 
-            // Actualizamos los datos de session
-            req.session.userLogged = {
-                ...userLogged,
-                ...userEdits,
-                avatar: req.file ? req.file.filename : userLogged.avatar,
-            };
+                if (req.file && userLogged.avatar != 'default.jpg') {
+                    fs.unlinkSync('./public/images/users/' + userLogged.avatar);
+                }
+                // Actualizamos los datos de session
+                req.session.userLogged = {
+                    ...userLogged,
+                    ...userEdits,
+                    avatar: req.file ? req.file.filename : userLogged.avatar,
+                };
+                
+                return res.redirect('/');
+            } catch (error) {
+                throw new Error(
+                    'Ocurrio un problema al actualizar la informacion, intente nuevamente mas tarde'
+                );
+            }
         }
-        return res.redirect('/');
     },
     confirmDelete: async (req, res) => {
         const userToDelete = await db.User.findOne({
@@ -167,17 +216,27 @@ const usersController = {
     },
     delete: async (req, res) => {
         const userLogged = req.session.userLogged;
-        // Borramos el avatar guardado
-        fs.unlinkSync('./public/images/users/' + userLogged.avatar);
         // Borramos al user de la BD
-        await db.User.destroy({
-            where: {
-                id: parseInt(req.params.id),
-            },
-        });
+        try {
+            await db.User.destroy({
+                where: {
+                    id: parseInt(req.params.id),
+                },
+            });
+            // Borramos el avatar guardado, solo si no tiene asiganada la imagen por defecto. Sino, la borraria
+            if (userLogged.avatar != 'default.jpg') {
+                fs.unlinkSync('./public/images/users/' + userLogged.avatar);
+            }
+            req.session.destroy();
+            res.clearCookie('userEmail');
+            return res.redirect('/');
+        } catch (error) {
+            throw new Error(
+                'Ocurrio un error al intentar eliminar la cuenta, por favor intente nuevamente'
+            );
+        }
+
         // Destruimos la session
-        req.session.destroy();
-        return res.redirect('/');
     },
 };
 module.exports = usersController;
